Changes abundance values from integer to float to allow for relative abundances to be used for sequence and bin abundances. #3 #4

+++ src/phylotree.cpp
@@ -11,7 +11,7 @@ PhyloTree::PhyloTree(){
 	addSeqToTree("unknown;", 1);
 }
 /******************************************************************************/
-int PhyloTree::addSeqToTree(string seqTaxonomy, int numReps){
+int PhyloTree::addSeqToTree(string seqTaxonomy, float numReps){
 
     numSeqs += numReps;
